pyproject.toml
[project]
name = "gymapp"
version = "1.0.0"
description = "A FastAPI-based gym assistant chatbot with conversation memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.2",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "openai>=2.6.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[project.scripts]
gym-chatbot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
